单元测试

+++ app-call-center/src/main/java/com/lsxy/call/center/dao/AgentSkillDao.java
@@ -16,10 +16,10 @@ public interface AgentSkillDao extends BaseDaoInterface<AgentSkill, Serializable
 
     @Modifying
     @Query("delete from AgentSkill where agent=:agent")
-    public long deleteByAgent(@Param("agent") String agent);
+    public int deleteByAgent(@Param("agent") String agent);
 
     @Modifying
     @Query("update AgentSkill a set active=:active where agent=:agent and name=:name")
-    public long updateActiveByAgent(@Param("active") Integer active,@Param("agent") String agent,@Param("name") String name);
+    public int updateActiveByAgent(@Param("active") Integer active,@Param("agent") String agent,@Param("name") String name);
 
 }
